Changed router. Routes now grouped by request type
Routes are now group by request type (GET, POST, PUT, DELETE).
Previously they were group by area (Subjects, Topics, Lessons, etc).

Only routes corresponding to the HTTP method in use are added to the routes array now.
So when doing a GET request, routes for GET requests are added to the routes array, but routes for POST requests are not.
Hopefully this saves a little bit of memory usage.