Reworked authorization system. Now uses JWT
Authorization system now works using JWT.

Signing into the api works mostly the same, except now a session cookie is not created.
Instead a JSON object is returned containing
- The JWT
- The expiry time (in JWT also, but is for convience)

In order to verify and do things that require authorization, the user must send the JWT as a request header name 'idToken' (accessed via $_SERVER{'HTTP_IDTOKEN'])

This token is then validated to check:
- The JWT is in valid format
- The issuer is correct
- The audience is correct
- The JWT has not expired
- The JWT has not been modified

If all the check passed, the user id is returned and the user is validated to do the action. (such as create a resource)