Updated controller/model code for tests, testquestions. Fixed bug with modify existing.
Updated queries in model files.
Cleaned up code in controller file since some checks are now redundant.

Fixed bug with modify existing for subjects, topics, lessons, tests, testquestions.
Did not check if the new values were unique for fields that must be unique.

For example:
A topic exists, with subject_id and name 'a'.
2 topic cannot have the same name if they have the same subject_id values.
2 topics can have the same name if they have different subject_id values.

Previously the modify exist code would let you change the name to 'a'.
This would cause an update error in SQL. It would not change anything, but on the code-side it was caught as a 500 error.
It has now been correctly updated so that it checks if a record with the new name and subject_id exists. If it does, the api returns 400 and message stating the name is not unique.