Subjects nows supports hidden attribute. Removed homepageContent
Removed homepageContent as it will no longer be used.

Subject now supports the 'hidden' attribute. When getting subjects it is included as either true/false.
Subjects now return the number of topics within the subject as 'topicCount'.

Added static function to App.php. boolToString.
Takes a boolean value. Returns '1' if true, '0' if false.
Returns null if given value is null or not boolean.