Added groups controller and model. Updated test data for testing groups.
Added groups controller and model.
Handles groups and managing users in groups.
Added most if not all functionality.

Added:
GET groups
Gets groups that the current user is a member of.
GET group/all
Gets all groups (admin only)
GET groups/:groupid
Gets group with specified id.
GET groups/:groupid/members
Gets users who are a member of the specified group.
GET groups/:groupid/nonmembers
Gets users who are not a member of the specified group.

POST groups
Creates a new group, based on given  content POST param, with the current user becoming the 1st and only member.
POST groups/:groupid
Modifies an existing group. Must be a member of the group or an admin.
DELETE groups/:groupid
Deletes an existing group. Must be amember of the group or an admin.

POST groups/:groupid/members/:userid
Adds a specified user to the specified group. Must be member of group or admin to use. Specified user must not already be a member of the group.
DELETE groups/:groupid/members/:userid.
Deletes a specified user from the specified group. Must be member of group or admin to use. Specified user must already be a member of the group.