Added support for subject_admins.
Added support for subject_admins.
Can get all subject admins, for a specific subject, based on the given count and offset values. User must be signed in to do this, however, admin is not required.
GET subjects/:sid/admins

Can add yourself as a subject_admin for a subject. User must be an admin to do this. Will return 400 if you are already a subject_admin for the subject.
POST subjects/:sid/admins

Can remove yourself as a subject_admin for a subject. You must be an admin to do this. Will return 400 if you are not an admin for the specified subject.
DELETE subjects/:sid/admins

Added event to database schema that runs everyday at 00:00:00.
Deletes all subject_admins records where the corresponding user is not an admin (does not have privilegeLevel_id value of 2).

Added some test data for subject_admins to testData.sql. Includes some non-admin subject_subjects, for testing the new database event.