Added route for getting whether current user is a subject admin for a specific subject.
Added hacky solution for getting whether current user is a subject admin for a specific subject.
Could not find effective alternative to this.
Must be admin to use, or will get 401. Must be give subject that exists, or will get 404.
Returns JSON object containing 1 attribute 'isSubjectAdmin', which a boolean.
Route is:
GET subjects/:sid/admins/me