Fixed bug where new subjects/topics/lessons/tests would not be returned.
Adding better support for the hidden attribute resulted in the creation of a method like getContentByIDAdmin, which acted like the old getContentByID method, not taking the hidden attribute into account.

The create/modify methods for subjects, topics, lessons, and tests were not updated to use this new method. This resulted in the frontend not being able to switch to the new resource when it was created, as the created resource would always be auto-hidden (had 0 children) and so the getContentByID method would not return it.