Bugfix - Modify test now includes hidden when checking if attributes modify. Small update to lesson formatRecords func.
Modify test now checks given hidden value when seeing if any values are given.
Previously it did not, so if hidden was the only value being changed, it would return 400, stating that no values were given to be changed.
If any other values were given, hidden would update correctly.
Fixed by including hidden in the check, so that if hidden and only hidden is changed, the modify request continues.

Updated ```formatRecords($records)``` method of ```Lessons.php```. Previously set body to addslashed(body), as it was HTML. Now it just returns body, as the frontend automatically adds \ before things that need escaping.