[Testing] - Fixed bug when attempting to authorize using Google with an invalid id_token
## Bug:
When attempting to authorize using a Google account, if an invalid id_token was sent, a UnexpectedValueException would be thrown (and not caught). This resulted in the user seeing the exception (in the test environment at least) and getting HTTP code 200.
They did not receive an authorization token however.

By invalid token, I mean a token that is not in JWT format (not a.b.c). The code acted correctly when sending a token that was in a valid format (but did not verify the user).

---

## Fix:
Surrounded API call with a try catch statement.
If an UnexpectedValueExpection is thrown, ```$payload``` is set to ```null```, resulting in
```
if (!isset($payload)) {
    http_response_code(400); return '`' . Auth::GOOGLE_TOKEN_NAME . '` is not valid.';
}
```
being triggered.