[Testing] - Fixed issue with being able to access 'child content' inside of hidden content.
Was a bug where non-admins could access content inside of hidden content.
E.g. access a visible topic inside of a hidden subject. This is wrong as if a subject is hidden, all of the content contained should also be hidden.

Getting topics (by subject / by id)
Updated query to check that the parent subject is not hidden.

Getting lessons (by topic [and therefore subject] / by id)
Updated query to check that:
- The parent topic is not hidden.
- The grandparent subject is not hidden.